+++ sdk/communication/Azure.Communication.Chat/tests/ChatClients/ChatClientsLiveTests.cs
@@ -7,9 +7,9 @@ using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
 #region Snippet:Azure_Communication_Chat_Tests_E2E_UsingStatements
-using Azure.Communication.Administration;
-using Azure.Communication.Administration.Models;
+using Azure.Core;
 using Azure.Communication;
+using Azure.Communication.Identity;
 //@@ using Azure.Communication.Chat;
 #endregion Snippet:Azure_Communication_Chat_Tests_E2E_UsingStatements
 
@@ -472,18 +472,18 @@ namespace Azure.Communication.Chat.Tests
         {
             Response<CommunicationUserIdentifier> threadMember = communicationIdentityClient.CreateUser();
             IEnumerable<CommunicationTokenScope> scopes = new[] { CommunicationTokenScope.Chat };
-            Response<CommunicationUserToken> tokenResponseThreadMember = communicationIdentityClient.IssueToken(threadMember.Value, scopes);
+            Response<AccessToken> tokenResponseThreadMember = communicationIdentityClient.IssueToken(threadMember.Value, scopes);
 
-            return (tokenResponseThreadMember.Value.User, tokenResponseThreadMember.Value.Token);
+            return (threadMember.Value, tokenResponseThreadMember.Value.Token);
         }
 
         private async Task<(CommunicationUserIdentifier user, string token)> CreateUserAndTokenAsync(CommunicationIdentityClient communicationIdentityClient)
         {
             Response<CommunicationUserIdentifier> threadMember = await communicationIdentityClient.CreateUserAsync();
             IEnumerable<CommunicationTokenScope> scopes = new[] { CommunicationTokenScope.Chat };
-            Response<CommunicationUserToken> tokenResponseThreadMember = await communicationIdentityClient.IssueTokenAsync(threadMember.Value, scopes);
+            Response<AccessToken> tokenResponseThreadMember = await communicationIdentityClient.IssueTokenAsync(threadMember.Value, scopes);
 
-            return (tokenResponseThreadMember.Value.User, tokenResponseThreadMember.Value.Token);
+            return (threadMember.Value, tokenResponseThreadMember.Value.Token);
         }
     }
 }