+++ sdk/formrecognizer/Azure.AI.FormRecognizer/src/FormLine.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 namespace Azure.AI.FormRecognizer.Models
 {
     /// <summary>
-    /// Represents a line of text recognized from the input document.
+    /// Represents a line of text  and its appearance.
     /// </summary>
     public class FormLine : FormElement
     {
@@ -35,6 +35,11 @@ namespace Azure.AI.FormRecognizer.Models
         /// </summary>
         public IReadOnlyList<FormWord> Words { get; }
 
+        /// <summary>
+        /// An object representing the appearance of the text line.
+        /// </summary>
+        public TextAppearance Appearance { get; }
+
         private static IReadOnlyList<FormWord> ConvertWords(IReadOnlyList<TextWord> textWords, int pageNumber)
         {
             List<FormWord> rawWords = new List<FormWord>();
@@ -46,8 +51,5 @@ namespace Azure.AI.FormRecognizer.Models
 
             return rawWords;
         }
-
-        /// <summary> Text appearance properties. </summary>
-        public Appearance Appearance { get; }
     }
 }