+++ sdk/servicebus/Microsoft.Azure.WebJobs.Extensions.ServiceBus/tests/ServiceBusTriggerAttributeTests.cs
@@ -1,38 +1,38 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-using Xunit;
+using NUnit.Framework;
 
 namespace Microsoft.Azure.WebJobs.ServiceBus.UnitTests
 {
     public class ServiceBusTriggerAttributeTests
     {
-        [Fact]
+        [Test]
         public void Constructor_Queue_SetsExpectedValues()
         {
             ServiceBusTriggerAttribute attribute = new ServiceBusTriggerAttribute("testqueue");
-            Assert.Equal("testqueue", attribute.QueueName);
+            Assert.AreEqual("testqueue", attribute.QueueName);
             Assert.Null(attribute.SubscriptionName);
             Assert.Null(attribute.TopicName);
 
             attribute = new ServiceBusTriggerAttribute("testqueue");
-            Assert.Equal("testqueue", attribute.QueueName);
+            Assert.AreEqual("testqueue", attribute.QueueName);
             Assert.Null(attribute.SubscriptionName);
             Assert.Null(attribute.TopicName);
         }
 
-        [Fact]
+        [Test]
         public void Constructor_Topic_SetsExpectedValues()
         {
             ServiceBusTriggerAttribute attribute = new ServiceBusTriggerAttribute("testtopic", "testsubscription");
             Assert.Null(attribute.QueueName);
-            Assert.Equal("testtopic", attribute.TopicName);
-            Assert.Equal("testsubscription", attribute.SubscriptionName);
+            Assert.AreEqual("testtopic", attribute.TopicName);
+            Assert.AreEqual("testsubscription", attribute.SubscriptionName);
 
             attribute = new ServiceBusTriggerAttribute("testtopic", "testsubscription");
             Assert.Null(attribute.QueueName);
-            Assert.Equal("testtopic", attribute.TopicName);
-            Assert.Equal("testsubscription", attribute.SubscriptionName);
+            Assert.AreEqual("testtopic", attribute.TopicName);
+            Assert.AreEqual("testsubscription", attribute.SubscriptionName);
         }
     }
 }